initial implementation of multi-broadcasting support
