+++ src/broadcasters/index.ts
@@ -1,6 +1,2 @@
-import lambda from "./lambda-broadcaster";
-import { Broadcaster } from "../types";
-
-export default {
-  lambda,
-} as { [name: string]: Broadcaster };
+export * from "./HttpBroadcaster";
+export * from "./Broadcaster";
